+++ src/main/java/dms/dms/controller/MemberController.java
@@ -5,7 +5,6 @@ import dms.dms.service.MemberService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -34,14 +33,13 @@ public class MemberController {
 
 
     @PostMapping("/member/login")
-    public String login(@ModelAttribute MemberDTO memberDTO, HttpSession session, Model model){
+    public String login(@ModelAttribute MemberDTO memberDTO, HttpSession session){
         System.out.println("memberController.login");
         System.out.println("memberDTO = "+memberDTO);
         MemberDTO loginResult = memberService.login(memberDTO);
         if (loginResult != null) {
             // login 성공
-            session.setAttribute("loginId", loginResult.getMemberId());
-            model.addAttribute("memberID", memberDTO.getMemberId());
+            session.setAttribute("memberId", loginResult.getMemberId());
             return "redirect:/main";
         } else {
             // login 실패
