+++ src/main/java/dms/dms/controller/StudyController.java
@@ -16,6 +16,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
@@ -38,7 +39,7 @@ public class StudyController {
     }
 
     @GetMapping(value="/study/studyHome")
-    public String studyListByMemberID(@RequestParam String memberId, Model model, @PageableDefault(page = 0, size = 10, sort="id", direction = Sort.Direction.ASC) Pageable pageable) { // 홈 화면 띄우기
+    public String studyListByMemberID(@SessionAttribute(name = "memberId", required = false) String memberId, Model model, @PageableDefault(page = 0, size = 3, sort="id", direction = Sort.Direction.ASC) Pageable pageable) { // 홈 화면 띄우기
 
         Page<Study> list = studyService.findStudiesByMemberID(memberId, pageable);
         int nowPage = list.getPageable().getPageNumber()+1;
@@ -53,31 +54,14 @@ public class StudyController {
         return "/study/studyHome";
     }
 
-    // 목록 보기 제목순 정렬
-//    @GetMapping(value="/study/studyHome/sortByTitle")
-//    public String studyListSortByTitle(Model model, @PageableDefault(page = 0, size = 10, sort="title", direction = Sort.Direction.ASC) Pageable pageable) { // 홈 화면 띄우기
-//
-//        Page<Study> list = studyService.findStudies(pageable);
-//        int nowPage = list.getPageable().getPageNumber()+1;
-//        int startPage = Math.max(nowPage-4,1);
-//        int endPage = Math.min(nowPage + 5, list.getTotalPages());
-//
-//        model.addAttribute("studys", list);
-//        model.addAttribute("nowPage", nowPage);
-//        model.addAttribute("startPage", startPage);
-//        model.addAttribute("endPage", endPage);
-//
-//        return "/study/studyHome";
-//    }
-
     @GetMapping(value="/study/studyCreate")
-    public String studyInsert(@RequestParam String memberId, Model model) { // 추가 화면 띄우기
+    public String studyInsert(@SessionAttribute(name = "memberId", required = false) String memberId, Model model) { // 추가 화면 띄우기
         model.addAttribute("memberID", memberId);
         return "/study/studyCreate";
     }
 
     @PostMapping(value="/study/studyCreates")
-    public String insert(@RequestParam String memberId, Study studyInfo, MultipartFile files) throws IOException { // 공부 기록 추가 메소드
+    public String insert(@SessionAttribute(name = "memberId", required = false) String memberId, Study studyInfo, MultipartFile files) throws IOException { // 공부 기록 추가 메소드
 
         Study study = new Study();
         System.out.println(memberId);
@@ -110,7 +94,7 @@ public class StudyController {
 
         studyService.saveStudy(study);
 
-        return "redirect:/study/studyHome?memberId="+memberId;
+        return "redirect:/study/studyHome";
     }
 
     @GetMapping(value="/study/studyContent")
@@ -172,13 +156,24 @@ public class StudyController {
         study.setTitle(studyInfo.getTitle());
         study.setContent(studyInfo.getContent());
         study.setDate(studyInfo.getDate());
-        study.setFilesavename(studyInfo.getFilesavename());
-        study.setFilesavepath(studyInfo.getFilesavepath());
-        study.setFileoriginname(studyInfo.getFileoriginname());
+
+        String originalName = studyInfo.getFileoriginname();
+        if(!originalName.isEmpty()) {
+            String uuid = UUID.randomUUID().toString();
+            //String extension = originalName.substring(originalName.lastIndexOf("."));
+            String saveName = uuid + originalName;
+            String savePath = fileDir + saveName;
+            study.setFileoriginname(originalName);
+            study.setFilesavename(saveName);
+            study.setFilesavepath(savePath);
+
+            File saveFile = new File(savePath);
+            files.transferTo(saveFile);
+        }
 
         studyService.saveStudy(study);
 
-        return "redirect:/study/studyHome?memberId="+study.getMemberId();
+        return "redirect:/study/studyHome";
     }
 
 }
