+++ src/main/resources/application.properties
@@ -1,6 +1,8 @@
 file.dir = C:/Users/user/dms_file
 # ?? ????? ??1
 
+server.servlet.session.tracking-modes=cookie
+
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://127.0.0.1:3306/dms
 useSSL=false
