--- src/main/resources/static/index.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="header.css">
-    <link rel="stylesheet" href="login.css">
-    <link rel="stylesheet" href="footer.css">
-    <title>Welcome to DMS</title>
-</head>
-<body>
-<header id="header">
-    <div class="inner">
-        <a href="index.html" id="dms"><img src="img/DMS.png" height="80" onclick=""></a>
-        <ul class="nav">
-            <li><a href="board.html" class="nav_item">게시판</a></li>
-            <li><a href="/study/studyHome" class="nav_item">공부기록</a></li>
-            <li><a href="plan.html" class="nav_item">일정</a></li>
-        </ul>
-        <span id="messege"><img src="img/messege.png" id = messege_icon onclick=""></span>
-        <span id="profile">
-                   <li><a href="#"><img src="img/profile.png" id = profile_icon  onclick=""></a>
-                       <ul id="profile_dropdown">
-                          <li><a href="#" class="dropdown_item">내 정보</a></li>
-                          <li><a href="#" class="dropdown_item">로그아웃</a></li>
-                       </ul>
-                 </li>
-            </span>
-    </div>
-</header>
-
-
-<section id="contents">
-    <div class="inner">
-        <div id="contents_container">
-            <div class="calendar">달력</div>
-            <div class="login">
-                <div id="login_inner">
-                    <form id="login_form" action="/member/login" method="post">
-                        <ul class="login_item">
-                            <li><input name="memberEmail" id="ID" class="item" type="text" placeholder="아이디를 입력하세요."></li>
-                            <li><input name="memberPassword" id="PW" class="item" type="password" placeholder="비밀번호를 입력하세요."></li>
-                            <input id="login_button" class="item" type="submit" value="로그인">
-                            <a href="/member/save">회원가입</a>
-                            <a href="#" id="find_id_pw">아이디/비밀번호 찾기</a>
-
-                        </ul>
-
-                    </form>
-                </div>
-            </div>
-            <div class="playing">공모전</div>
-            <div class="board">게시판</div>
-        </div>
-    </div>
-</section>
-<footer id="footer">
-    <div class="inner">
-        <div id="footer_container">
-            footer
-        </div>
-    </div>
-
-</footer>
-</body>
-</html>
