+++ src/main/resources/templates/main.html
@@ -1,18 +1,5 @@
-
 <!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>main</title>
-</head>
-<body>
-session 값 확인 : <p th:text="${session.loginId}" style="color: red"></p>
-</body>
-</html>
-
-
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html lang="en">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -23,17 +10,16 @@ session 값 확인 : <p th:text="${session.loginId}" style="color: red"></p>
     <title>Welcome to DMS</title>
 </head>
 <body>
-    <header id="header">
-        <h2>로그인 한 MAIN</h2>
-        <div class="inner">
-            <a href="main" id="dms"><img src="/img/DMS.png" height="80" onclick=""></a>
-            <ul class="nav">
-                 <li><a href="/schedule/plan" class="nav_item">게시판</a></li>
-                 <li><a th:href="@{/study/studyHome(memberId=${session.loginId})}" class="nav_item">공부기록</a></li>
-                 <li><a th:href="@{/schedule/plan(memberID=${session.loginId})}" class="nav_item">일정</a></li>
-            </ul>
-             <span id="message"><a href="/message"><img src="/img/message.png" id = message_icon onclick=""></a></span>
-             <span id="profile">
+<header id="header">
+    <div class="inner">
+        <a href="main" id="dms"><img src="/img/DMS.png" height="80" onclick=""></a>
+        <ul class="nav">
+            <li><a href="board.html" class="nav_item">게시판</a></li>
+            <li><a href="/study/studyHome" class="nav_item">공부기록</a></li>
+            <li><a href="/schedule/plan" class="nav_item">일정</a></li>
+        </ul>
+        <span id="message"><a href="/message"><img src="/img/message.png" id = message_icon onclick=""></a></span>
+        <span id="profile">
                    <li><a href="#"><img src="/img/profile.png" id = profile_icon  onclick=""></a>
                        <ul id="profile_dropdown">
                           <li><a href="#" class="dropdown_item">내 정보</a></li>
@@ -41,25 +27,25 @@ session 값 확인 : <p th:text="${session.loginId}" style="color: red"></p>
                        </ul>
                  </li>
             </span>
+    </div>
+</header>
+<section id="contents">
+    <div class="inner">
+        <div id="contents_container">
+            <div class="detail">디테일 일정</div>
+            <div class="calendar">달력</div>
+            <div class="playing">공모전</div>
+            <div class="board">게시판</div>
         </div>
-    </header>
-    <section id="contents">
-        <div class="inner">
-            <div id="contents_container">
-                <div class="detail">디테일 일정</div>
-                <div class="calendar">달력</div>
-                <div class="playing">공모전</div>
-                <div class="board">게시판</div>
-            </div>
+    </div>
+</section>
+<footer id="footer">
+    <div class="inner">
+        <div id="footer_container">
+            footer
         </div>
-    </section>
-    <footer id="footer">
-        <div class="inner">
-            <div id="footer_container">
-                footer
-            </div>
-        </div>
+    </div>
 
-    </footer>
+</footer>
 </body>
 </html>
