+++ src/main/resources/templates/study/studyCreate.html
@@ -29,8 +29,6 @@
                         <input type="file" name="files" onchange="selectFile(this);" />
                     </label>
                 </div>
-                <button type="button" onclick="removeFile(this);" class="btns del_btn"><span>삭제</span></button>
-                <button type="button" onclick="addFile();" class="btns fn_add_btn"><span>파일 추가</span></button>
             </div>
         </div>
         <script>
@@ -40,13 +38,7 @@
                 const file = element.files[0];
                 const filename = element.closest('.file_input').firstElementChild;
 
-                // 1. 파일 선택 창에서 취소 버튼이 클릭된 경우
-                if ( !file ) {
-                    filename.value = '';
-                    return false;
-                }
-
-                // 2. 파일 크기가 10MB를 초과하는 경우
+                // 파일 크기가 10MB를 초과하는 경우
                 const fileSize = Math.floor(file.size / 1024 / 1024);
                 if (fileSize > 10) {
                     alert('10MB 이하의 파일로 업로드해 주세요.');
@@ -55,42 +47,13 @@
                     return false;
                 }
 
-                // 3. 파일명 지정
+                // 파일명 지정
                 filename.value = file.name;
             }
-
-
-            // 파일 추가
-            function addFile() {
-                const fileDiv = document.createElement('div');
-                fileDiv.innerHTML =`
-            <div class="file_input">
-                <input type="text" readonly />
-                <label> 첨부파일
-                    <input type="file" name="files" onchange="selectFile(this);" />
-                </label>
-            </div>
-            <button type="button" onclick="removeFile(this);" class="btns del_btn"><span>삭제</span></button>
-        `;
-                document.querySelector('.file_list').appendChild(fileDiv);
-            }
-
-
-            // 파일 삭제
-            function removeFile(element) {
-                const fileAddBtn = element.nextElementSibling;
-                if (fileAddBtn) {
-                    const inputs = element.previousElementSibling.querySelectorAll('input');
-                    inputs.forEach(input => input.value = '')
-                    return false;
-                }
-                element.parentElement.remove();
-            }
-
         </script>
         <button type="submit">등록</button>
     </form>
-</div> <!-- /container -->
+</div>
 
 </body>
 </html>
